Retire SkillsBench line budget pins
