+++ examples/control_plane/repo-python-line-budget-smoke.py
@@ -16,11 +16,15 @@ DEFAULT_MAX_LINES = 2000
 LEGACY_OVERSIZED_RETIREMENT_PLANS = {
     "examples/skillsbench-benchmark-run-smoke.py": (
         "Extract SkillsBench ledger/reduce-only fixture groups into narrower "
-        "smokes, then lower this pin back below 14808."
+        "smokes, then lower this pin again after the next stable split."
+    ),
+    "loopx/benchmark_adapters/skillsbench.py": (
+        "Continue extracting SkillsBench discovery, counter, and reducer helpers "
+        "into focused adapter modules, then lower this pin again."
     ),
     "scripts/skillsbench_automation_loop.py": (
         "Extract native app-server Goal/reduce-only closeout helpers into "
-        "scripts or benchmark adapter modules, then lower this pin below 16791."
+        "scripts or benchmark adapter modules, then lower this pin again."
     ),
 }
 
@@ -28,7 +32,7 @@ LEGACY_OVERSIZED_LIMITS = {
     "examples/benchmark-run-ledger-smoke.py": 2106,
     "examples/control_plane/quota-plan-smoke.py": 2176,
     "examples/skillsbench-app-server-goal-worker-smoke.py": 3056,
-    "examples/skillsbench-benchmark-run-smoke.py": 14900,
+    "examples/skillsbench-benchmark-run-smoke.py": 14430,
     "examples/skillsbench-host-local-launch-plan-smoke.py": 2373,
     "examples/control_plane/status-markdown-smoke.py": 2607,
     "examples/terminal-bench-harbor-runner-ingest-smoke.py": 2759,
@@ -36,7 +40,7 @@ LEGACY_OVERSIZED_LIMITS = {
     "loopx/benchmark.py": 2875,
     "loopx/benchmark_adapters/agentissue.py": 2644,
     "loopx/benchmark_adapters/agents_last_exam.py": 3998,
-    "loopx/benchmark_adapters/skillsbench.py": 5943,
+    "loopx/benchmark_adapters/skillsbench.py": 5817,
     "examples/control_plane/work-lane-contract-smoke.py": 2336,
     "loopx/benchmark_adapters/skillsbench_acp_relay.py": 3144,
     "loopx/benchmark_adapters/terminal_bench.py": 10045,
@@ -49,7 +53,7 @@ LEGACY_OVERSIZED_LIMITS = {
     "loopx/terminal_bench_agent.py": 2056,
     "loopx/todos.py": 2105,
     "scripts/harbor_host_codex_goal_agent.py": 2140,
-    "scripts/skillsbench_automation_loop.py": 16855,
+    "scripts/skillsbench_automation_loop.py": 16682,
 }
 
 
@@ -90,6 +94,7 @@ def main() -> None:
         if limit > DEFAULT_MAX_LINES
         and path in {
             "examples/skillsbench-benchmark-run-smoke.py",
+            "loopx/benchmark_adapters/skillsbench.py",
             "scripts/skillsbench_automation_loop.py",
         }
         and not LEGACY_OVERSIZED_RETIREMENT_PLANS.get(path)
